+++ modules/core/23-commitment/types/merkle_test.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"testing"
 
-	abci "github.com/line/ostracon/abci/types"
 	"github.com/stretchr/testify/require"
+	abci "github.com/tendermint/tendermint/abci/types"
 
 	"github.com/line/ibc-go/v3/modules/core/23-commitment/types"
 )