+++ modules/core/23-commitment/types/utils_test.go
@@ -3,9 +3,9 @@ package types_test
 import (
 	"fmt"
 
-	abci "github.com/line/ostracon/abci/types"
-	crypto "github.com/line/ostracon/proto/ostracon/crypto"
 	"github.com/stretchr/testify/require"
+	abci "github.com/tendermint/tendermint/abci/types"
+	"github.com/tendermint/tendermint/proto/tendermint/crypto"
 
 	"github.com/line/ibc-go/v3/modules/core/23-commitment/types"
 )