+++ modules/light-clients/99-ostracon/types/update_test.go
@@ -22,8 +22,6 @@ func (suite *OstraconTestSuite) TestCheckHeaderAndUpdateState() {
 		newHeader       *types.Header
 		currentTime     time.Time
 		bothValSet      *octypes.ValidatorSet
-		voterSet        *octypes.VoterSet
-		bothVoterSet    *octypes.VoterSet
 		signers         []octypes.PrivValidator
 		bothSigners     []octypes.PrivValidator
 	)
@@ -44,7 +42,6 @@ func (suite *OstraconTestSuite) TestCheckHeaderAndUpdateState() {
 	altVal := ibctesting.NewTestValidator(altPubKey, revisionHeight)
 	// Create alternative validator set with only altVal, invalid update (too much change in valSet)
 	altValSet := octypes.NewValidatorSet([]*octypes.Validator{altVal})
-	altVoterSet := octypes.WrapValidatorsToVoterSet(altValSet.Validators)
 	altSigners := []octypes.PrivValidator{altPrivVal}
 
 	testCases := []struct {
@@ -58,7 +55,7 @@ func (suite *OstraconTestSuite) TestCheckHeaderAndUpdateState() {
 			setup: func(suite *OstraconTestSuite) {
 				clientState = types.NewClientState(chainID, types.DefaultTrustLevel, trustingPeriod, ubdPeriod, maxClockDrift, height, commitmenttypes.GetSDKSpecs(), upgradePath, false, false)
 				consensusState = types.NewConsensusState(suite.clientTime, commitmenttypes.NewMerkleRoot(suite.header.Header.GetAppHash()), suite.valsHash)
-				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus1.RevisionHeight), height, suite.headerTime, suite.valSet, suite.valSet, voterSet, voterSet, signers)
+				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus1.RevisionHeight), height, suite.headerTime, suite.valSet, suite.valSet, signers)
 				currentTime = suite.now
 			},
 			expFrozen: false,
@@ -69,7 +66,7 @@ func (suite *OstraconTestSuite) TestCheckHeaderAndUpdateState() {
 			setup: func(suite *OstraconTestSuite) {
 				clientState = types.NewClientState(chainID, types.DefaultTrustLevel, trustingPeriod, ubdPeriod, maxClockDrift, height, commitmenttypes.GetSDKSpecs(), upgradePath, false, false)
 				consensusState = types.NewConsensusState(suite.clientTime, commitmenttypes.NewMerkleRoot(suite.header.Header.GetAppHash()), suite.valsHash)
-				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus5.RevisionHeight), height, suite.headerTime, bothValSet, suite.valSet, bothVoterSet, voterSet, bothSigners)
+				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus5.RevisionHeight), height, suite.headerTime, bothValSet, suite.valSet, bothSigners)
 				currentTime = suite.now
 			},
 			expFrozen: false,
@@ -80,7 +77,7 @@ func (suite *OstraconTestSuite) TestCheckHeaderAndUpdateState() {
 			setup: func(suite *OstraconTestSuite) {
 				clientState = types.NewClientState(chainID, types.DefaultTrustLevel, trustingPeriod, ubdPeriod, maxClockDrift, height, commitmenttypes.GetSDKSpecs(), upgradePath, false, false)
 				consensusState = types.NewConsensusState(suite.clientTime, commitmenttypes.NewMerkleRoot(suite.header.Header.GetAppHash()), bothValSet.Hash())
-				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus1.RevisionHeight), height, suite.headerTime, bothValSet, bothValSet, bothVoterSet, bothVoterSet, bothSigners)
+				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus1.RevisionHeight), height, suite.headerTime, bothValSet, bothValSet, bothSigners)
 				currentTime = suite.now
 			},
 			expFrozen: false,
@@ -92,7 +89,7 @@ func (suite *OstraconTestSuite) TestCheckHeaderAndUpdateState() {
 				clientState = types.NewClientState(chainID, types.DefaultTrustLevel, trustingPeriod, ubdPeriod, maxClockDrift, height, commitmenttypes.GetSDKSpecs(), upgradePath, false, false)
 				consensusState = types.NewConsensusState(suite.clientTime, commitmenttypes.NewMerkleRoot(suite.header.Header.GetAppHash()), suite.valsHash)
 				consStateHeight = heightMinus3
-				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightMinus1.RevisionHeight), heightMinus3, suite.headerTime, bothValSet, suite.valSet, bothVoterSet, voterSet, bothSigners)
+				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightMinus1.RevisionHeight), heightMinus3, suite.headerTime, bothValSet, suite.valSet, bothSigners)
 				currentTime = suite.now
 			},
 			expFrozen: false,
@@ -104,7 +101,7 @@ func (suite *OstraconTestSuite) TestCheckHeaderAndUpdateState() {
 				clientState = types.NewClientState(chainIDRevision1, types.DefaultTrustLevel, trustingPeriod, ubdPeriod, maxClockDrift, height, commitmenttypes.GetSDKSpecs(), upgradePath, false, false)
 				consensusState = types.NewConsensusState(suite.clientTime, commitmenttypes.NewMerkleRoot(suite.header.Header.GetAppHash()), suite.valsHash)
 				consStateHeight = heightMinus3
-				newHeader = suite.chainA.CreateOCClientHeader(chainIDRevision0, int64(height.RevisionHeight), heightMinus3, suite.headerTime, bothValSet, suite.valSet, bothVoterSet, voterSet, bothSigners)
+				newHeader = suite.chainA.CreateOCClientHeader(chainIDRevision0, int64(height.RevisionHeight), heightMinus3, suite.headerTime, bothValSet, suite.valSet, bothSigners)
 				currentTime = suite.now
 			},
 			expPass: true,
@@ -114,7 +111,7 @@ func (suite *OstraconTestSuite) TestCheckHeaderAndUpdateState() {
 			setup: func(suite *OstraconTestSuite) {
 				clientState = types.NewClientState(chainID, types.DefaultTrustLevel, trustingPeriod, ubdPeriod, maxClockDrift, heightPlus1, commitmenttypes.GetSDKSpecs(), upgradePath, false, false)
 				consensusState = types.NewConsensusState(suite.clientTime, commitmenttypes.NewMerkleRoot(suite.header.Header.GetAppHash()), suite.valsHash)
-				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus1.RevisionHeight), height, suite.headerTime, suite.valSet, suite.valSet, voterSet, voterSet, signers)
+				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus1.RevisionHeight), height, suite.headerTime, suite.valSet, suite.valSet, signers)
 				currentTime = suite.now
 				ctx := suite.chainA.GetContext().WithBlockTime(currentTime)
 				// Store the header's consensus state in client store before UpdateClient call
@@ -128,7 +125,7 @@ func (suite *OstraconTestSuite) TestCheckHeaderAndUpdateState() {
 			setup: func(suite *OstraconTestSuite) {
 				clientState = types.NewClientState(chainID, types.DefaultTrustLevel, trustingPeriod, ubdPeriod, maxClockDrift, heightPlus1, commitmenttypes.GetSDKSpecs(), upgradePath, false, false)
 				consensusState = types.NewConsensusState(suite.clientTime, commitmenttypes.NewMerkleRoot(suite.header.Header.GetAppHash()), suite.valsHash)
-				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus1.RevisionHeight), height, suite.headerTime, suite.valSet, suite.valSet, voterSet, voterSet, signers)
+				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus1.RevisionHeight), height, suite.headerTime, suite.valSet, suite.valSet, signers)
 				currentTime = suite.now
 				ctx := suite.chainA.GetContext().WithBlockTime(currentTime)
 				// Change the consensus state of header and store in client store to create a conflict
@@ -146,7 +143,7 @@ func (suite *OstraconTestSuite) TestCheckHeaderAndUpdateState() {
 				// create an intermediate consensus state with the same time as the newHeader to create a time violation.
 				// header time is after client time
 				consensusState = types.NewConsensusState(suite.clientTime, commitmenttypes.NewMerkleRoot(suite.header.Header.GetAppHash()), suite.valsHash)
-				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus5.RevisionHeight), height, suite.headerTime, suite.valSet, suite.valSet, voterSet, voterSet, signers)
+				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus5.RevisionHeight), height, suite.headerTime, suite.valSet, suite.valSet, signers)
 				currentTime = suite.now
 				prevConsensusState := types.NewConsensusState(suite.headerTime, commitmenttypes.NewMerkleRoot(suite.header.Header.GetAppHash()), suite.valsHash)
 				ctx := suite.chainA.GetContext().WithBlockTime(currentTime)
@@ -164,7 +161,7 @@ func (suite *OstraconTestSuite) TestCheckHeaderAndUpdateState() {
 				// create the next consensus state with the same time as the intermediate newHeader to create a time violation.
 				// header time is after clientTime
 				consensusState = types.NewConsensusState(suite.clientTime, commitmenttypes.NewMerkleRoot(suite.header.Header.GetAppHash()), suite.valsHash)
-				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus1.RevisionHeight), height, suite.headerTime, suite.valSet, suite.valSet, voterSet, voterSet, signers)
+				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus1.RevisionHeight), height, suite.headerTime, suite.valSet, suite.valSet, signers)
 				currentTime = suite.now
 				nextConsensusState := types.NewConsensusState(suite.headerTime, commitmenttypes.NewMerkleRoot(suite.header.Header.GetAppHash()), suite.valsHash)
 				ctx := suite.chainA.GetContext().WithBlockTime(currentTime)
@@ -180,7 +177,7 @@ func (suite *OstraconTestSuite) TestCheckHeaderAndUpdateState() {
 			setup: func(suite *OstraconTestSuite) {
 				clientState = types.NewClientState(chainID, types.DefaultTrustLevel, trustingPeriod, ubdPeriod, maxClockDrift, height, commitmenttypes.GetSDKSpecs(), upgradePath, false, false)
 				consensusState = types.NewConsensusState(suite.clientTime, commitmenttypes.NewMerkleRoot(suite.header.Header.GetAppHash()), suite.valsHash)
-				newHeader = suite.chainA.CreateOCClientHeader("ethermint", int64(heightPlus1.RevisionHeight), height, suite.headerTime, suite.valSet, suite.valSet, voterSet, voterSet, signers)
+				newHeader = suite.chainA.CreateOCClientHeader("ethermint", int64(heightPlus1.RevisionHeight), height, suite.headerTime, suite.valSet, suite.valSet, signers)
 				currentTime = suite.now
 			},
 			expFrozen: false,
@@ -191,7 +188,7 @@ func (suite *OstraconTestSuite) TestCheckHeaderAndUpdateState() {
 			setup: func(suite *OstraconTestSuite) {
 				clientState = types.NewClientState(chainIDRevision0, types.DefaultTrustLevel, trustingPeriod, ubdPeriod, maxClockDrift, height, commitmenttypes.GetSDKSpecs(), upgradePath, false, false)
 				consensusState = types.NewConsensusState(suite.clientTime, commitmenttypes.NewMerkleRoot(suite.header.Header.GetAppHash()), suite.valsHash)
-				newHeader = suite.chainA.CreateOCClientHeader(chainIDRevision1, 1, height, suite.headerTime, suite.valSet, suite.valSet, voterSet, voterSet, signers)
+				newHeader = suite.chainA.CreateOCClientHeader(chainIDRevision1, 1, height, suite.headerTime, suite.valSet, suite.valSet, signers)
 				currentTime = suite.now
 			},
 			expPass: false,
@@ -201,7 +198,7 @@ func (suite *OstraconTestSuite) TestCheckHeaderAndUpdateState() {
 			setup: func(suite *OstraconTestSuite) {
 				clientState = types.NewClientState(chainIDRevision1, types.DefaultTrustLevel, trustingPeriod, ubdPeriod, maxClockDrift, clienttypes.NewHeight(1, 1), commitmenttypes.GetSDKSpecs(), upgradePath, false, false)
 				consensusState = types.NewConsensusState(suite.clientTime, commitmenttypes.NewMerkleRoot(suite.header.Header.GetAppHash()), suite.valsHash)
-				newHeader = suite.chainA.CreateOCClientHeader(chainIDRevision1, 3, height, suite.headerTime, suite.valSet, suite.valSet, voterSet, voterSet, signers)
+				newHeader = suite.chainA.CreateOCClientHeader(chainIDRevision1, 3, height, suite.headerTime, suite.valSet, suite.valSet, signers)
 				currentTime = suite.now
 			},
 			expFrozen: false,
@@ -212,7 +209,7 @@ func (suite *OstraconTestSuite) TestCheckHeaderAndUpdateState() {
 			setup: func(suite *OstraconTestSuite) {
 				clientState = types.NewClientState(chainID, types.DefaultTrustLevel, trustingPeriod, ubdPeriod, maxClockDrift, height, commitmenttypes.GetSDKSpecs(), upgradePath, false, false)
 				consensusState = types.NewConsensusState(suite.clientTime, commitmenttypes.NewMerkleRoot(suite.header.Header.GetAppHash()), suite.valsHash)
-				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus1.RevisionHeight), height, suite.headerTime, bothValSet, suite.valSet, bothVoterSet, voterSet, bothSigners)
+				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus1.RevisionHeight), height, suite.headerTime, bothValSet, suite.valSet, bothSigners)
 				currentTime = suite.now
 			},
 			expFrozen: false,
@@ -223,7 +220,7 @@ func (suite *OstraconTestSuite) TestCheckHeaderAndUpdateState() {
 			setup: func(suite *OstraconTestSuite) {
 				clientState = types.NewClientState(chainID, types.DefaultTrustLevel, trustingPeriod, ubdPeriod, maxClockDrift, height, commitmenttypes.GetSDKSpecs(), upgradePath, false, false)
 				consensusState = types.NewConsensusState(suite.clientTime, commitmenttypes.NewMerkleRoot(suite.header.Header.GetAppHash()), bothValSet.Hash())
-				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus1.RevisionHeight), height, suite.headerTime, suite.valSet, bothValSet, voterSet, bothVoterSet, signers)
+				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus1.RevisionHeight), height, suite.headerTime, suite.valSet, bothValSet, signers)
 				currentTime = suite.now
 			},
 			expFrozen: false,
@@ -234,7 +231,7 @@ func (suite *OstraconTestSuite) TestCheckHeaderAndUpdateState() {
 			setup: func(suite *OstraconTestSuite) {
 				clientState = types.NewClientState(chainID, types.DefaultTrustLevel, trustingPeriod, ubdPeriod, maxClockDrift, height, commitmenttypes.GetSDKSpecs(), upgradePath, false, false)
 				consensusState = types.NewConsensusState(suite.clientTime, commitmenttypes.NewMerkleRoot(suite.header.Header.GetAppHash()), suite.valsHash)
-				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus5.RevisionHeight), height, suite.headerTime, altValSet, suite.valSet, altVoterSet, voterSet, altSigners)
+				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus5.RevisionHeight), height, suite.headerTime, altValSet, suite.valSet, altSigners)
 				currentTime = suite.now
 			},
 			expFrozen: false,
@@ -245,7 +242,7 @@ func (suite *OstraconTestSuite) TestCheckHeaderAndUpdateState() {
 			setup: func(suite *OstraconTestSuite) {
 				clientState = types.NewClientState(chainID, types.DefaultTrustLevel, trustingPeriod, ubdPeriod, maxClockDrift, height, commitmenttypes.GetSDKSpecs(), upgradePath, false, false)
 				consensusState = types.NewConsensusState(suite.clientTime, commitmenttypes.NewMerkleRoot(suite.header.Header.GetAppHash()), suite.valsHash)
-				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus5.RevisionHeight), height, suite.headerTime, bothValSet, bothValSet, bothVoterSet, bothVoterSet, bothSigners)
+				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus5.RevisionHeight), height, suite.headerTime, bothValSet, bothValSet, bothSigners)
 				currentTime = suite.now
 			},
 			expFrozen: false,
@@ -256,7 +253,7 @@ func (suite *OstraconTestSuite) TestCheckHeaderAndUpdateState() {
 			setup: func(suite *OstraconTestSuite) {
 				clientState = types.NewClientState(chainID, types.DefaultTrustLevel, trustingPeriod, ubdPeriod, maxClockDrift, height, commitmenttypes.GetSDKSpecs(), upgradePath, false, false)
 				consensusState = types.NewConsensusState(suite.clientTime, commitmenttypes.NewMerkleRoot(suite.header.Header.GetAppHash()), suite.valsHash)
-				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus1.RevisionHeight), height, suite.headerTime, suite.valSet, suite.valSet, voterSet, voterSet, signers)
+				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus1.RevisionHeight), height, suite.headerTime, suite.valSet, suite.valSet, signers)
 				// make current time pass trusting period from last timestamp on clientstate
 				currentTime = suite.now.Add(trustingPeriod)
 			},
@@ -268,7 +265,7 @@ func (suite *OstraconTestSuite) TestCheckHeaderAndUpdateState() {
 			setup: func(suite *OstraconTestSuite) {
 				clientState = types.NewClientState(chainID, types.DefaultTrustLevel, trustingPeriod, ubdPeriod, maxClockDrift, height, commitmenttypes.GetSDKSpecs(), upgradePath, false, false)
 				consensusState = types.NewConsensusState(suite.clientTime, commitmenttypes.NewMerkleRoot(suite.header.Header.GetAppHash()), suite.valsHash)
-				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus1.RevisionHeight), height, suite.now.Add(time.Minute), suite.valSet, suite.valSet, voterSet, voterSet, signers)
+				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus1.RevisionHeight), height, suite.now.Add(time.Minute), suite.valSet, suite.valSet, signers)
 				currentTime = suite.now
 			},
 			expFrozen: false,
@@ -279,7 +276,7 @@ func (suite *OstraconTestSuite) TestCheckHeaderAndUpdateState() {
 			setup: func(suite *OstraconTestSuite) {
 				clientState = types.NewClientState(chainID, types.DefaultTrustLevel, trustingPeriod, ubdPeriod, maxClockDrift, height, commitmenttypes.GetSDKSpecs(), upgradePath, false, false)
 				consensusState = types.NewConsensusState(suite.clientTime, commitmenttypes.NewMerkleRoot(suite.header.Header.GetAppHash()), suite.valsHash)
-				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus1.RevisionHeight), height, suite.clientTime, suite.valSet, suite.valSet, voterSet, voterSet, signers)
+				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus1.RevisionHeight), height, suite.clientTime, suite.valSet, suite.valSet, signers)
 				currentTime = suite.now
 			},
 			expFrozen: false,
@@ -290,7 +287,7 @@ func (suite *OstraconTestSuite) TestCheckHeaderAndUpdateState() {
 			setup: func(suite *OstraconTestSuite) {
 				clientState = types.NewClientState(chainID, types.DefaultTrustLevel, trustingPeriod, ubdPeriod, maxClockDrift, height, commitmenttypes.GetSDKSpecs(), upgradePath, false, false)
 				consensusState = types.NewConsensusState(suite.clientTime, commitmenttypes.NewMerkleRoot(suite.header.Header.GetAppHash()), suite.valsHash)
-				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus1.RevisionHeight), height, suite.headerTime, suite.valSet, suite.valSet, voterSet, voterSet, signers)
+				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightPlus1.RevisionHeight), height, suite.headerTime, suite.valSet, suite.valSet, signers)
 				// cause new header to fail validatebasic by changing commit height to mismatch header height
 				newHeader.SignedHeader.Commit.Height = revisionHeight - 1
 				currentTime = suite.now
@@ -304,7 +301,7 @@ func (suite *OstraconTestSuite) TestCheckHeaderAndUpdateState() {
 				clientState = types.NewClientState(chainID, types.DefaultTrustLevel, trustingPeriod, ubdPeriod, maxClockDrift, clienttypes.NewHeight(height.RevisionNumber, heightPlus5.RevisionHeight), commitmenttypes.GetSDKSpecs(), upgradePath, false, false)
 				consensusState = types.NewConsensusState(suite.clientTime, commitmenttypes.NewMerkleRoot(suite.header.Header.GetAppHash()), suite.valsHash)
 				// Make new header at height less than latest client state
-				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightMinus1.RevisionHeight), height, suite.headerTime, suite.valSet, suite.valSet, voterSet, voterSet, signers)
+				newHeader = suite.chainA.CreateOCClientHeader(chainID, int64(heightMinus1.RevisionHeight), height, suite.headerTime, suite.valSet, suite.valSet, signers)
 				currentTime = suite.now
 			},
 			expFrozen: false,
@@ -323,9 +320,6 @@ func (suite *OstraconTestSuite) TestCheckHeaderAndUpdateState() {
 			// Create signer array and ensure it is in same order as bothValSet
 			_, suiteVal := suite.valSet.GetByIndex(0)
 			bothSigners = ibctesting.CreateSortedSignerArray(altPrivVal, suite.privVal, altVal, suiteVal)
-
-			voterSet = octypes.WrapValidatorsToVoterSet(suite.valSet.Validators)
-			bothVoterSet = octypes.WrapValidatorsToVoterSet(bothValSet.Validators)
 
 			consStateHeight = height // must be explicitly changed
 			// setup test